requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.21.0
scipy>=1.7.0

# Data handling and storage
datasets>=2.14.0
pyyaml>=6.0
json5>=0.9.0

# Async and concurrency
asyncio-coroutine-types>=0.0.1
aiofiles>=0.8.0

# Mathematical operations
einops>=0.7.0
scikit-learn>=1.0.0

# Optional GPU acceleration  
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Model fine-tuning (optional)
peft>=0.6.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Documentation
markdown>=3.4.0

# Utilities
tqdm>=4.64.0
colorama>=0.4.0
python-dateutil>=2.8.0

# Type hints
typing-extensions>=4.0.0

# Memory optimization
psutil>=5.8.0

# Networking (for distributed processing)
requests>=2.28.0
aiohttp>=3.8.0

# Configuration management
omegaconf>=2.3.0

# Monitoring and logging
tensorboard>=2.10.0

# Data visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Text processing
nltk>=3.7.0

# Embedding models (optional)
sentence-transformers>=2.2.0

# Audio processing (for future multimodal features)
librosa>=0.9.0
soundfile>=0.10.0

# Image processing (for future multimodal features)  
pillow>=9.0.0

# Security
cryptography>=37.0.0

# Legacy dependencies (maintaining compatibility)
sentencepiece>=0.1.97
protobuf>=3.20.0
flask>=2.0.0
